# openapi.yaml
openapi: 3.0.0
info:
    title: Sistema RC Staff API
    description: API REST frontend-agnostic e desacoplada em Go para o sistema RC Staff (musculação, corrida, RAG, SVED e relatórios administrativos).
    version: 0.1.0-rc
servers:
    - url: http://localhost:5000
      description: Servidor de desenvolvimento local
...
paths:
...
    /api/v1/feedback/{hash}:
        get:
            operationId: getFeedbackByHash
            summary: Verifica se há feedback preenchido para o link ativo (público)
            tags: [Feedback]
            parameters:
                - name: hash
                  in: path
                  required: true
                  schema:
                    type: string
                    example: a1b2c3d4e5f6
                  description: Hash público do link da ficha ou plano de corrida. Usado em `/api/v1/feedback/{hash}`.
            responses:
                '200':
                    description: Status do feedback
            description: Verifica se há feedback preenchido para o link ativo (público) (`GET /api/v1/feedback/{hash}`).
        post:
            operationId: postFeedbackByHash
            summary: Envia feedback qualitativo de treino (público)
            tags: [Feedback]
            parameters:
                - name: hash
                  in: path
                  required: true
                  schema:
                    type: string
                    example: a1b2c3d4e5f6
                  description: Hash público do link da ficha ou plano de corrida. Usado em `/api/v1/feedback/{hash}`.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [feedback]
                            properties:
                                feedback:
                                    type: string
                                    example: Treino concluído sem dor.
                        example:
                            feedback: Treino concluído sem dor.
                description: Corpo da requisição para `POST /api/v1/feedback/{hash}`.
            responses:
                '200':
                    description: Feedback registrado
            description: Envia feedback qualitativo de treino (público) (`POST /api/v1/feedback/{hash}`).
...
components:
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 'JWT obtido em POST /api/v1/auth/login. Enviar header Authorization: Bearer <token>.'
...